refactor(proxy): move `pollBroker()` to `Start()`
...from `runSession()`.
`runSession()` sounds like it should only be called
when we actually get a client.

This slightly modifies the behavior of the proxy:
now it will not wait for the client connection to succeed / fail
before polling the broker another time.
This might make the verbose logs a little more messy.
But the poll interval is still respected.